refactor: reorganize and fix imports to accommodate previous changes
