--- components/Colors.js
@@ -1,12 +1,0 @@
-const Colors = {
-  0: "#FFFFFF",
-  100: "#F4F4F5",
-  200: "#E4E4E7",
-  300: "#D4D4D8",
-  400: "#A1A1AA",
-  500: "#71717A",
-  600: "#27272A",
-  700: "#18181B",
-  800: "#09090B",
-};
-export default Colors;
